New xmltv indexing scheme

+++ dune_plugin/lib/epg_manager.php
@@ -27,16 +27,16 @@ class Epg_Manager
     protected $xmltv_urls;
 
     /**
-     * contains filename of the cached xmltv file (if original is packed points to unpacked version)
-     * @var array
-     */
-    protected $xml_cached_file;
-
-    /**
-     * contains parsed epg for current xmltv file (ALL data!)
+     * contains parsed epg for channel
      * @var array
      */
     public $xmltv_data;
+
+    /**
+     * contains index for current xmltv file
+     * @var array
+     */
+    public $xmltv_index;
 
     /**
      * @param default_config $config
@@ -49,13 +49,6 @@ class Epg_Manager
     public function clear_xmltv_urls()
     {
         $this->xmltv_urls = array();
-        $this->xml_cached_file = array();
-    }
-
-    public function clear_epg_memory_cache()
-    {
-        $this->epg_cache = array();
-        $this->xmltv_data = null;
     }
 
     /**
@@ -65,16 +58,16 @@ class Epg_Manager
     public function set_xmltv_url($idx, $xmltv_url)
     {
         $this->xmltv_urls[$idx] = $xmltv_url;
-        $this->xml_cached_file[$idx] = $this->config->plugin_info['app_name'] . "_" . hash('crc32', $xmltv_url) . ".xmltv";
-        hd_print(__METHOD__ . ": index: $idx, set cached file: {$this->xml_cached_file[$idx]} for url {$this->xmltv_urls[$idx]}");
+        hd_print(__METHOD__ . ": index: $idx, url {$this->xmltv_urls[$idx]}");
     }
 
     /**
-     * @param $idx string
+     * @param $plugin_cookies
      * @return string
      */
-    public function get_xmltv_url($idx)
+    public function get_xmltv_url($plugin_cookies)
     {
+        $idx = $this->get_xmltv_idx($plugin_cookies);
         return (isset($this->xmltv_urls[$idx])) ? $this->xmltv_urls[$idx] : '';
     }
 
@@ -87,6 +80,7 @@ class Epg_Manager
     }
 
     /**
+     * @param $plugin_cookies
      * @return string
      */
     public function get_xmltv_idx($plugin_cookies)
@@ -121,13 +115,41 @@ class Epg_Manager
     }
 
     /**
-     * @param $idx string
      * @param $plugin_cookies
      * @return string
      */
-    public function get_xml_cached_file($idx, $plugin_cookies)
+    public function get_xml_cached_filename($plugin_cookies)
     {
-        return isset($this->xml_cached_file[$idx]) ? self::get_xcache_dir($plugin_cookies) . $this->xml_cached_file[$idx] : '';
+        $idx = $this->get_xmltv_idx($plugin_cookies);
+        return isset($this->xmltv_urls[$idx]) ? hash('crc32', $this->xmltv_urls[$idx]) : '';
+    }
+
+    /**
+     * @param $plugin_cookies
+     * @return boolean
+     */
+    public function is_xml_cache_set($plugin_cookies)
+    {
+        $name = $this->get_xml_cached_filename($plugin_cookies);
+        return !empty($name);
+    }
+
+    /**
+     * @param $plugin_cookies
+     * @return string
+     */
+    public function get_xml_cached_file($plugin_cookies)
+    {
+        return self::get_xcache_dir($plugin_cookies) . $this->get_xml_cached_filename($plugin_cookies) . '.xmltv';
+    }
+
+    /**
+     * @param $plugin_cookies
+     * @return string
+     */
+    public function get_xml_cached_file_index($plugin_cookies)
+    {
+        return self::get_xcache_dir($plugin_cookies) . $this->get_xml_cached_filename($plugin_cookies) . '.index';
     }
 
     /**
@@ -141,7 +163,9 @@ class Epg_Manager
     public function get_day_epg_items(Channel $channel, $day_start_ts, &$plugin_cookies)
     {
         try {
-            $epg_source = isset($plugin_cookies->epg_source) ? $plugin_cookies->epg_source : SetupControlSwitchDefs::switch_epg1;
+            $epg_source = isset($plugin_cookies->{Starnet_Epg_Setup_Screen::SETUP_ACTION_EPG_SOURCE})
+                ? $plugin_cookies->{Starnet_Epg_Setup_Screen::SETUP_ACTION_EPG_SOURCE}
+                : SetupControlSwitchDefs::switch_epg1;
             $epg_id = $this->config->get_epg_id($channel, $epg_source);
             if (empty($epg_id)) {
                 throw new Exception("EPG ID for $epg_source source not defined");
@@ -154,7 +178,7 @@ class Epg_Manager
 
             $params = $this->config->get_epg_params($epg_source);
             if ($epg_source === Plugin_Constants::EPG_INTERNAL) {
-                $params[Epg_Params::EPG_DOMAIN] = $this->get_xmltv_url($this->get_xmltv_idx($plugin_cookies));
+                $params[Epg_Params::EPG_DOMAIN] = $this->get_xmltv_url($plugin_cookies);
             } else if (empty($params[Epg_Params::EPG_URL])) {
                 $epg_source = ($epg_source === Plugin_Constants::EPG_FIRST) ? Plugin_Constants::EPG_SECOND : Plugin_Constants::EPG_FIRST;
                 $params = $this->config->get_epg_params($epg_source);
@@ -209,12 +233,14 @@ class Epg_Manager
             $program_epg = array();
             if ($epg_source === Plugin_Constants::EPG_INTERNAL) {
                 hd_print(__METHOD__ . ": Fetching EPG ID: '$epg_id' from xmltv");
-                $program_epg = $this->get_epg_xmltv($epg_id, $this->get_xmltv_idx($plugin_cookies), $plugin_cookies);
+                $program_epg = $this->get_epg_xmltv($epg_id, $plugin_cookies);
             } else {
                 $from_cache = false;
                 if (file_exists($epg_cache_file)) {
                     $now = time();
-                    $max_check_time = 3600 * 24 * (isset($plugin_cookies->epg_cache_ttl) ? $plugin_cookies->epg_cache_ttl : 3);
+                    $max_check_time = 3600 * 24 * (isset($plugin_cookies->{Starnet_Epg_Setup_Screen::SETUP_ACTION_EPG_CACHE_TTL})
+                            ? $plugin_cookies->{Starnet_Epg_Setup_Screen::SETUP_ACTION_EPG_CACHE_TTL}
+                            : 3);
                     $cache_expired = filemtime($epg_cache_file) + $max_check_time;
                     if ($cache_expired > time()) {
                         $program_epg = self::load_cache($epg_cache_file);
@@ -292,30 +318,44 @@ class Epg_Manager
 
     /**
      * @param $id string
-     * @param $idx
      * @param $plugin_cookies
      * @return false|array
      */
-    public function get_epg_xmltv($id, $idx, $plugin_cookies)
+    public function get_epg_xmltv($id, $plugin_cookies)
     {
         try {
+            if (!$this->is_xml_cache_set($plugin_cookies)) {
+                throw new Exception("xmltv file is not set");
+            }
 
-            $cached_file = $this->get_xml_cached_file($idx, $plugin_cookies);
-            if (empty($cached_file)) {
-                throw new Exception("xmltv file for index: $idx is not set");
+            if (empty($this->xmltv_index)) {
+                $index_file = $this->get_xml_cached_file_index($plugin_cookies);
+                hd_print(__METHOD__ . ": load index from file '$index_file'");
+                $data = json_decode(file_get_contents($index_file), true);
+                if (false !== $data) {
+                    $this->xmltv_index = $data;
+                }
+
+                if (empty($this->xmltv_index)) {
+                    throw new Exception("load index failed '$index_file'");
+                }
             }
 
             if (!isset($this->xmltv_data[$id])) {
-                $channel_idx = self::get_xcache_dir($plugin_cookies) . $id . "_" . basename($cached_file) . ".index";
-                if (!file_exists($channel_idx)) {
-                    throw new Exception("channel $id not found");
+                if (!isset($this->xmltv_index[$id])) {
+                    throw new Exception("xmltv index for channel $id is not exist");
                 }
 
-                $this->xmltv_data[$id] = json_decode(file_get_contents($channel_idx), true);
+                $channel_index = self::get_xcache_dir($plugin_cookies) . $this->xmltv_index[$id];
+                hd_print(__METHOD__ . "Check channel $id index: $channel_index");
+                if (!file_exists($channel_index)) {
+                    throw new Exception("index for channel $id not found: $channel_index");
+                }
+
+                $this->xmltv_data[$id] = json_decode(file_get_contents($channel_index), true);
             }
 
-
-            $file_object = self::open_xmltv_file($cached_file);
+            $file_object = self::open_xmltv_file($this->get_xml_cached_file($plugin_cookies));
             $ch_epg = array();
             foreach ($this->xmltv_data[$id] as $pos) {
                 $xml_str = '';
@@ -348,47 +388,42 @@ class Epg_Manager
     }
 
     /**
-     * @param $xml_cached_path
-     * @param $max_check_time
-     * @return boolean
+     * @param $plugin_cookies
+     * @return string if downloaded xmltv file exists it empty, otherwise contain error message
      */
-    public static function is_xmltv_cache_valid($xml_cached_path, $max_check_time)
+    public function is_xmltv_cache_valid($plugin_cookies)
     {
-        hd_print(__METHOD__ . ": Cached file path: $xml_cached_path");
-        if (!file_exists($xml_cached_path)) {
-            hd_print(__METHOD__ . ": Cached file path not exist");
-            return false;
-        }
-
-        $check_time_file = filemtime($xml_cached_path);
-        if ($check_time_file && $check_time_file + $max_check_time > time()) {
-            hd_print(__METHOD__ . ": Cached file index: $xml_cached_path not expired " . date("Y-m-d H:s", $check_time_file));
-            return true;
-        }
-        unlink($xml_cached_path);
-
-        return false;
-    }
-
-    /**
-     * @param $xmltv_url string
-     * @param $xml_cached_path string
-     * @return boolean|string
-     */
-    public static function download_xmltv_url($xmltv_url, $xml_cached_path)
-    {
-        if (file_exists($xml_cached_path)) {
-            return true;
-        }
-
-        hd_print(__METHOD__ . ": Storage space in cache dir: " . HD::get_storage_size(dirname($xml_cached_path)));
-
         try {
+            if (!$this->is_xml_cache_set($plugin_cookies)) {
+                throw new Exception("Cached xmltv file is not set");
+            }
+
+            $cached_file = $this->get_xml_cached_file($plugin_cookies);
+            $cached_file_index = $this->get_xml_cached_file_index($plugin_cookies);
+            hd_print(__METHOD__ . ": Checking cached xmltv file: $cached_file");
+            if (!file_exists($cached_file)) {
+                hd_print(__METHOD__ . ": Cached xmltv file not exist");
+            } else {
+                $check_time_file = filemtime($cached_file);
+                $max_cache_time = 3600 * 24 * (isset($plugin_cookies->{Starnet_Epg_Setup_Screen::SETUP_ACTION_EPG_CACHE_TTL})
+                        ? $plugin_cookies->{Starnet_Epg_Setup_Screen::SETUP_ACTION_EPG_CACHE_TTL}
+                        : 3);
+                if ($check_time_file && $check_time_file + $max_cache_time > time()) {
+                    hd_print(__METHOD__ . ": Cached file: $cached_file is not expired " . date("Y-m-d H:s", $check_time_file));
+                    return '';
+                }
+
+                hd_print(__METHOD__ . ": clear cached file: $cached_file");
+                unlink($cached_file);
+            }
+
+            $xmltv_url = $this->get_xmltv_url($plugin_cookies);
             if (empty($xmltv_url)) {
                 throw new Exception("XMTLV EPG url not set");
             }
 
-            $tmp_filename = get_temp_path(basename($xmltv_url));
+            hd_print(__METHOD__ . ": Storage space in cache dir: " . HD::get_storage_size(dirname($cached_file)));
+            $tmp_filename = $cached_file . 'tmp';
             $last_mod_file = HD::http_save_document($xmltv_url, $tmp_filename);
             hd_print(__METHOD__ . ": Last changed time on server: " . date("Y-m-d H:s", $last_mod_file));
 
@@ -399,9 +434,9 @@ class Epg_Manager
                     throw new Exception("Failed to open $tmp_filename");
                 }
 
-                $dest = fopen($xml_cached_path, 'wb');
+                $dest = fopen($cached_file, 'wb');
                 if (!$dest) {
-                    throw new Exception("Failed to open $xml_cached_path");
+                    throw new Exception("Failed to open $cached_file");
                 }
 
                 $res = stream_copy_to_stream($gz, $dest);
@@ -409,41 +444,69 @@ class Epg_Manager
                 fclose($dest);
                 unlink($tmp_filename);
                 if ($res === false) {
-                    throw new Exception("Failed to unpack $tmp_filename to $xml_cached_path");
+                    throw new Exception("Failed to unpack $tmp_filename to $cached_file");
                 }
-                hd_print(__METHOD__ . ": $res bytes written to $xml_cached_path");
+                hd_print(__METHOD__ . ": $res bytes written to $cached_file");
             } else {
-                rename($tmp_filename, $xml_cached_path);
+                if (file_exists($cached_file)) {
+                    unlink($cached_file);
+                }
+                rename($tmp_filename, $cached_file);
+            }
+
+            if (file_exists($cached_file_index)) {
+                hd_print(__METHOD__ . ": clear cached file index: $cached_file_index");
+                unlink($cached_file_index);
             }
         } catch (Exception $ex) {
             hd_print(__METHOD__ . ": " . $ex->getMessage());
             return $ex->getMessage();
         }
 
-        return true;
+        return '';
     }
 
     /**
-     * @param $xml_cached_path string
+     * @param $plugin_cookies
      * @param $epg_ids array
      */
-    public static function index_xmltv_file($xml_cached_path, $epg_ids)
+    public function index_xmltv_file($plugin_cookies, $epg_ids)
     {
+        $res = $this->is_xmltv_cache_valid($plugin_cookies);
+        if (!empty($res)) {
+            hd_print(__METHOD__ . ": Error load xmltv: $res");
+            return;
+        }
+
+        $cached_file_index = $this->get_xml_cached_file_index($plugin_cookies);
+        if (file_exists($cached_file_index)) {
+            hd_print(__METHOD__ . ": Load cache file index: $cached_file_index");
+            $this->xmltv_index = json_decode(file_get_contents($cached_file_index), true);
+            return;
+        }
+
+        hd_print(__METHOD__ . ": Cached file index: $cached_file_index is not valid need reindex");
+
+        $parse_all = isset($plugin_cookies->{Starnet_Epg_Setup_Screen::SETUP_ACTION_EPG_PARSE_ALL})
+            && $plugin_cookies->{Starnet_Epg_Setup_Screen::SETUP_ACTION_EPG_PARSE_ALL} === SetupControlSwitchDefs::switch_on;
+
+        $cache_file = $this->get_xml_cached_filename($plugin_cookies);
+        $cache_dir = self::get_xcache_dir($plugin_cookies);
+        $lock_file = "$cache_dir$cache_file.lock";
         try {
-            $lock_file = $xml_cached_path . ".lock";
             if (file_exists($lock_file)) {
-                throw new Exception("File indexed now, skipped");
+                throw new Exception("File is indexed now, skipped");
             }
 
             file_put_contents($lock_file, '');
-            hd_print(__METHOD__ . ": Reindexing $xml_cached_path");
+
+            $cached_path = $this->get_xml_cached_file($plugin_cookies);
+            $file_object = self::open_xmltv_file($cached_path);
+            hd_print(__METHOD__ . ": Reindexing $cached_path");
             $t = microtime(1);
 
-            $count = 0;
-            $parent_name = basename($xml_cached_path);
-            $parent_path = dirname($xml_cached_path);
-            $file_object = self::open_xmltv_file($xml_cached_path);
             $prev_channel = '';
+            $xmltv_index = array();
             $xmltv_data = array();
             while (!$file_object->eof()) {
                 $pos = $file_object->ftell();
@@ -466,7 +529,7 @@ class Epg_Manager
                 $channel = substr($line, $ch_start, $ch_end - $ch_start);
 
                 // if epg id not present in current channels list skip it
-                if (!isset($epg_ids[$channel])) {
+                if (!$parse_all && !is_null($epg_ids) && !isset($epg_ids[$channel])) {
                     continue;
                 }
 
@@ -475,8 +538,9 @@ class Epg_Manager
                         $prev_channel = $channel;
                     } else {
                         if (!empty($xmltv_data)) {
-                            file_put_contents($parent_path . "/$prev_channel" . "_$parent_name.index", json_encode($xmltv_data));
-                            $count++;
+                            $index_name = $cache_file . "_" . hash("crc32", $prev_channel) . '.index';
+                            $xmltv_index[$prev_channel] = $index_name;
+                            file_put_contents("$cache_dir$index_name", json_encode($xmltv_data));
                         }
 
                         $prev_channel = $channel;
@@ -487,17 +551,25 @@ class Epg_Manager
             }
 
             if (!empty($prev_channel) && !empty($xmltv_data)) {
-                file_put_contents($parent_path . "/$prev_channel" . "_$parent_name.index", json_encode($xmltv_data));
+                $index_name = $cache_file . "_" . hash("crc32", $prev_channel) . '.index';
+                $xmltv_index[$prev_channel] = $index_name;
+                file_put_contents("$cache_dir$index_name", json_encode($xmltv_data));
+            }
+
+            if (!empty($xmltv_index)) {
+                hd_print(__METHOD__ . ": Save index: $cached_file_index");
+                file_put_contents($cached_file_index, json_encode($xmltv_index));
+                $this->xmltv_index = $xmltv_index;
             }
 
             hd_print(__METHOD__ . ": Reindexing XMLTV done: " . (microtime(1) - $t) . " secs");
-            hd_print(__METHOD__ . ": Total epg id's indexed: $count");
+            hd_print(__METHOD__ . ": Total epg id's indexed: " . count($xmltv_index));
         } catch (Exception $ex) {
             hd_print(__METHOD__ . ": " . $ex->getMessage());
         }
 
         HD::ShowMemoryUsage();
-        hd_print(__METHOD__ . ": Storage space in cache dir after reindexing: " . HD::get_storage_size(dirname($xml_cached_path)));
+        hd_print(__METHOD__ . ": Storage space in cache dir after reindexing: " . HD::get_storage_size($cache_dir));
 
         unlink($lock_file);
     }
@@ -510,6 +582,7 @@ class Epg_Manager
     {
         $this->epg_cache = array();
         $this->xmltv_data = null;
+        $this->xmltv_index = null;
 
         $dir = self::get_xcache_dir($plugin_cookies);
         hd_print(__METHOD__ . ": clear cache dir: $dir");
